FrSkyR9: initial telemetry support

+++ Multiprotocol/FrSkyDVX_common.ino
@@ -33,7 +33,7 @@ static uint16_t __attribute__((unused)) FrSkyX_CRCTable(uint8_t val)
 	val /= 16 ;
 	return word ^ (0x1081 * val) ;
 }
-uint16_t FrSkyX_crc(uint8_t *data, uint8_t len, uint8_t init=0)
+uint16_t FrSkyX_crc(uint8_t *data, uint8_t len, uint16_t init=0)
 {
 	uint16_t crc = init;
 	for(uint8_t i=0; i < len; i++)
